Sparql11Engine: Increase Basic Auth Header length
When setting an HTTP Authentication Header for a triple store
Sparql11Engine used a TextField. The default maximum length for a
TextField is 128. This turned out to cause a problem when the username
or password grows to long.

This commit increases the maximum length for the HTTP Authentication
Header Setting to 9999.