Add an AST Annotation Stage
Previously, the in-progress rewrite of the query engine consisted of
three stages: AST Generation, Query Plan Generation and Plan Execution.
This commit inserts a new stage between the AST and Query Plan
Generation Stages.

This stage annotates the AST with all relevant information (such as
relevant adapters and bundles) so that the query planner does not need
to lookup adapters but can just use the already searched information.