Properly deleting entity references in sub-bundles
This was caused by using a wrong key to look up the field value that has to be deleted for entity reference fields.
This resulted in an empty field value which in turn lead to just skipping the reference entirely.
I'm not sure why this only occurs in sub-bundles though.

Now we're looking up the field definitions to check if the current field is an entity reference and then set the lookup key to `target_id`.